update Makefile not to be limited to GCC
When the libraries are not built within the source tree of Asterisk, they are built via the Makefile. In that, the compiler was GCC, not allowing Clang, for example. Now, the Makefile uses the platform default compiler (GCC, Clang, or …). If GCC is not your default, but you want to restore GCC, go for something like `CC=gcc make` or change the default compiler system-wide via `sudo update-alternatives --config cc`.